debian/rules: Fix/Workaround for the newmat FTBFS (#812280)
What broke the build is that gcc 6 changed the default C++ standard
from C++98 to C++14.

Not all valid C++98 code is also valid C++11 and C++14 code.

Note that this just changed the default, when told to process C++98 code
gcc 6 does not differ in any significant way from gcc 5.

Making the code compatible with C++14 would be the best possible
solution, but as a workaround it is possible to fix the build with
this change to tell gcc that this is C++98 code.

CXXFLAGS set in debian/rules were not used at all, this is also fixed.